+++ internal/grpc/scheduled_function_rpc_test.go
@@ -70,7 +70,7 @@ func newTestEnvWithDispatch(t *testing.T) (*CloudEventsServiceImpl, *workflow.Ha
 	if err != nil {
 		t.Fatalf("token.NewSigner: %v", err)
 	}
-	dispatcher := NewProcessorDispatcher(registry, common.NewDefaultUUIDGenerator(), signer, "node-test", time.Minute)
+	dispatcher := NewProcessorDispatcher(registry, NewRoundRobinSelector(registry), common.NewDefaultUUIDGenerator(), signer, "node-test", time.Minute)
 
 	engine := workflow.NewEngine(factory, common.NewDefaultUUIDGenerator(), txMgr, workflow.WithExternalProcessing(dispatcher))
 	searchStore, _ := factory.AsyncSearchStore(context.Background())